Improve simulateSkillHelper file into three smaller helper files
