+++ src/askContainer/webViews/simulateSkillWebview.ts
@@ -6,12 +6,9 @@ import * as path from 'path';
 import { getSkillDetailsFromWorkspace } from '../../utils/skillHelper';
 import { DEFAULT_PROFILE, ERRORS, SIMULATOR_MESSAGE_TYPE } from '../../constants';
 import { loggableAskError } from '../../exceptions';
-import {
-    handleSkillStatusMessageFromWebview, handleLocaleMessageFromWebview,
-    handleUtteranceMessageFromWebview, handleExportMessageFromWebview,
-    handleActionMessageFromWebview, getReplayList, getNewViewPortMessage
-} from '../../utils/simulateSkillHelper';
 import { IViewport } from "apl-suggester";
+import * as simulateMessageHelper from '../../utils/simulateMessageHelper';
+import * as simulateNonSkillHelper from '../../utils/simulateNonSkillHelper';
 
 export class SimulateSkillWebview extends AbstractWebView {
     private loader: ViewLoader;
@@ -31,15 +28,15 @@ export class SimulateSkillWebview extends AbstractWebView {
     getHtmlForView(): string {
         Logger.debug(`Calling method: ${this.viewId}.getHtmlForView`);
 
-        const webview: vscode.Webview = this.getWebview() as vscode.Webview;
-        const simulateCss: vscode.Uri = webview.asWebviewUri(
+        //const webview: vscode.Webview = this.getWebview() as vscode.Webview;
+        const simulateCss: vscode.Uri = this.getWebview()?.asWebviewUri(
             vscode.Uri.file(
                 path.join(
                     this.extensionContext.extensionPath, 'media', 'simulate.css',
                 ),
             ));
 
-        const aplRenderUtils: vscode.Uri = webview.asWebviewUri(
+        const aplRenderUtils: vscode.Uri = this.getWebview()?.asWebviewUri(
             vscode.Uri.file(
                 path.join(
                     this.extensionContext.extensionPath, 'media/previewApl', 'aplRenderUtils.js',
@@ -61,7 +58,6 @@ export class SimulateSkillWebview extends AbstractWebView {
         });
     }
 
-
     async onReceiveMessageListener(message: Record<string, any>): Promise<void> {
         Logger.debug(`Calling method: ${this.viewId}.onReceiveMessageListener`);
 
@@ -71,27 +67,27 @@ export class SimulateSkillWebview extends AbstractWebView {
         const profile = Utils.getCachedProfile(this.extensionContext) ?? DEFAULT_PROFILE;
 
         if (message.type === SIMULATOR_MESSAGE_TYPE.SKILL_STATUS) {
-            const returnMessage: string | Record<string, any> = await handleSkillStatusMessageFromWebview(message, profile, skillId, this.extensionContext);
+            const returnMessage: string | Record<string, any> = await simulateMessageHelper.handleSkillStatusMessageFromWebview(message, profile, skillId, this.extensionContext);
             await this.getWebview()?.postMessage({
                 message: returnMessage,
                 type: SIMULATOR_MESSAGE_TYPE.SKILL_STATUS
             });
         }
         else if (message.type === SIMULATOR_MESSAGE_TYPE.LOCALE) {
-            const returnMessage: void | Record<string, any> = await handleLocaleMessageFromWebview(message, profile, skillId, this.extensionContext);
+            const returnMessage: void | Record<string, any> = await simulateMessageHelper.handleLocaleMessageFromWebview(message, profile, skillId, this.extensionContext);
             if (returnMessage !== null) {
                 await this.getWebview()?.postMessage(returnMessage);
             }
         }
         else if (message.type === SIMULATOR_MESSAGE_TYPE.UTTERANCE) {
-            const returnMessage = await handleUtteranceMessageFromWebview(message, profile, skillId, this.extensionContext);
+            const returnMessage = await simulateMessageHelper.handleUtteranceMessageFromWebview(message, profile, skillId, this.extensionContext);
             await this.getWebview()?.postMessage(returnMessage);
         }
         else if (message.type === SIMULATOR_MESSAGE_TYPE.EXPORT) {
-            handleExportMessageFromWebview(message, skillId, skillName, this.extensionContext);
+            void simulateMessageHelper.handleExportMessageFromWebview(message, skillId, skillName, this.extensionContext);
         }
         else if (message.type === SIMULATOR_MESSAGE_TYPE.ACTION) {
-            handleActionMessageFromWebview(message, skillId);
+            simulateMessageHelper.handleActionMessageFromWebview(message, skillId);
         }
         else {
             throw loggableAskError(ERRORS.UNRECOGNIZED_MESSAGE_FROM_WEBVIEW);
@@ -100,7 +96,7 @@ export class SimulateSkillWebview extends AbstractWebView {
 
     async replaySessionInSimulator(): Promise<void> {
         Logger.debug(`Calling method: ${this.viewId}.replaySessionInSimulator`);
-        const returnMessage = await getReplayList();
+        const returnMessage = await simulateNonSkillHelper.getReplayList();
         if (returnMessage !== null) {
             await this.getWebview()?.postMessage(returnMessage);
         }
@@ -109,7 +105,7 @@ export class SimulateSkillWebview extends AbstractWebView {
     //Get the document/datasource/new_viewport to change the preview.
     async changeViewport(viewport: IViewport): Promise<void> {
         Logger.verbose(`Calling method: ${this.viewId}.changeViewport, args: `, viewport.toString());
-        const returnMessage = getNewViewPortMessage(viewport);
+        const returnMessage = simulateNonSkillHelper.getNewViewPortMessage(viewport);
         await this.getWebview()?.postMessage(returnMessage);
     }
 }
